add `lflip`rflip options to mm
In q if one of the matrices being multiplied needs to be transposed,
such as in

q)A mmu flip B

the flip can be folded into the multiplication as

q)B mmu/:A

Besides avoiding extra memory allocation and work for the actual flip,
this changes the memory access pattern of the multiplication so that
both matrices are read sequentially (row by row), which is faster.

The new options allow a similar optimization with .qml.mm:

q).qml.mm[A] flip B

becomes

q).qml.mmx[`rflip;A;B]

See changes in example/ for some examples.

Suggested-by: Nick Psaris <nick.psaris@gmail.com>